fixed delete previous when none evaluated bug

+++ labeller/utils.py
@@ -1,8 +1,7 @@
-import re
-import random
 from .models import User
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
+import re
 
 def extract_entity(span, entity):
     locations = [m.start() for m in re.finditer(entity.lower(), span.lower())]
@@ -31,9 +30,10 @@ def validate_user(user_id):
 
 def delete_previous_label(Evaluated_Snippet, user, view):
     if view in ['spam', 'out_of_samples']:
-        previous_evaluated_snippet_id = max([s.id for s in Evaluated_Snippet.objects.all() if s.evaluator == user])
-        previous_snippet_id = Evaluated_Snippet.objects.get(id=previous_evaluated_snippet_id).snippet.id
-        Evaluated_Snippet.objects.get(id=previous_evaluated_snippet_id).delete()
+        previous_evaluated_snippet_ids = [s.id for s in Evaluated_Snippet.objects.all() if s.evaluator == user]
+        if len(previous_evaluated_snippet_ids) > 0:
+            previous_snippet_id = Evaluated_Snippet.objects.get(id=max(previous_evaluated_snippet_ids)).snippet.id
+            Evaluated_Snippet.objects.get(id=previous_evaluated_snippet_id).delete()
         return HttpResponseRedirect('/spam')
     elif view == 'opinion':
